working version of ecrecover
