+++ README.md
@@ -46,6 +46,8 @@ Note: Do not modify this section! It is auto-generated by `cobra` using `make ge
 
 - [polycli dumpblocks](doc/polycli_dumpblocks.md) - Export a range of blocks from a JSON-RPC endpoint.
 
+- [polycli ecrecover](doc/polycli_ecrecover.md) - Recovers and returns the public key of the signature
+
 - [polycli enr](doc/polycli_enr.md) - Convert between ENR and Enode format
 
 - [polycli fork](doc/polycli_fork.md) - Take a forked block and walk up the chain to do analysis.
